add page on legacy CopperheadOS branding
We'll want to fill in more details about the history of the project and
historical archives of the code, documentation, tweets, community, etc.

This will help us with trimming down the information on CopperheadOS in
the history page. We want to reduce emphasis on the ongoing conflict and
focus on informing people about what happened in the past.

Immense damage has been done to GrapheneOS by Copperhead's organized
misinformation campaign and we need to do our best to repair it.