sandboxed Play services compat layer is available
The plans to make a stub implementation have been shelved for the time
being since it wouldn't provide broad compatibility and the choice would
result in confusion. It's best for us to push for apps to function with
no Play services and we expect Windows 11 Android app support is going
to help a lot with this. If apps work with a stub implementation, then
they're perfectly capable of working without it available and can easily
fix the portability of their code. It may be something we decide to work
on in the future, but we need to see what happens in the next few years
and we now have a great way of handling it in the meantime.