Clarify that HTTP GET Requests must be used for the OpenSearch document to be parsed
Chromium-based browsers on Android only parse the OpenSearch document if GET HTTP requests are used with the query. Even if the website serves an OpenSearch document, and the user performs a search on the site, it will not be detected by the browser if it uses a POST HTTP request to handle the query.

Startpage is one example of a search engine which properly supports the OpenSearch standard, but uses POST HTTP requests to prevent searches showing up in the user's browsing history, and so Chromium-based browsers ignore the OpenSearch document.