start introducing conforming procedure style
The changes summarized:

* Use angle brackets (`>`) instead of arrow symbols (`➔`).
* Put a non-breaking space (`&#160;`) before each angle bracket.
* Wrap sequences and elements in bold element (`<b>`).
* Accessibility: Wrap angle bracket in a span tag with an aria-label.

References:

* https://developers.google.com/style/procedures#multi-action-procedures
* https://developers.google.com/style/ui-elements#angle-brackets

I use the entity number `&#160;` for the non-breaking space instead of
the named character entity `&nbsp;` because `xmllint` produces a linting
error (which is to be expected).

```
parser error : Entity 'nbsp' not defined
```

Ideally we want to use the named character entity (`&nbsp;`), but I
couldn't find a reasonable solution. In order not to block this change,
I've opted for the entity number (`&#160;`). If a reasonable solution is
found, the entity number can be easily replaced by the named character
entity.

The changes require more than just replacing the currently used arrow.
The surrounding context should be checked to determine whether a change
is appropriate at this point and whether it requires further changes in
the context itself.

The paragraphs are wrapped with the currently used 98 characters. If
using vim:

```
:set tw=98
```

Vim's wrapping is weird when applied several times. To work around this,
we can unwrap the paragraph first. To select a paragraph, unwrap and
rewrap: Point to the line of the paragraph, and type `vipJgqq` in normal
mode.