stop recommending OpenVPN, Orbot and RethinkDNS
OpenVPN is legacy technology. It's overly complex and has far too much
attack surface. It's a huge pile of sketchy legacy code. WireGuard is
what people should be using now. We were only recommending this due to
lack of WireGuard adoption in the past.

Orbot is a horribly maintained app full of memory corruption bugs and
poorly written code. It's not a trustworthy or safe implementation of
Tor for Android. Tor Project is working on a proper replacement which
does not appear to be available yet. We could make something ourselves
if others are failing to provide what's needed.

RethinkDNS was added as a recommendation in order to have an option for
people who want local filtering and other features while using a VPN. We
cannot recommend it anymore. We overlooked technical and other issues
because we wanted to give people an option to use. The bar was simply
that it was better than NetGuard. It's still temporarily mentioned as
an option supporting that combination of features but will be replaced
in the documentation as soon as possible.