avoid misinterpretation of update requirements
The update requirements exist because we need to include device support
code for each supported device including firmware, drivers and HALs. It
was easiest to define the requirement as having proper overall updates,
but we can clarify that we only need it for the device support code even
though it's not going to be updated by itself without the OS also being
updated. This adds a bit of additional text, but it avoids the current
situation where several people are misrepresenting this requirement.