use Protobuf library for Python from PyPI
Even with fairly up-to-date packages, many Linux distributions have not
recent enough Python Protobuf packages. It's better to use the PyPI
package, which is published by Google, rather than the distribution
packages.

Ideally, adevtool could handle this (with pinned package version and
hash, and hash-checking), but due to many users encountering this issue,
this change should be reasonable for now.