Explain how to fix WebUSB of Chromium on Ubuntu
The documentation previously claimed that snap-packaged chromium has WebUSB broken on Ubuntu. This is not entirely accurate as it is just disabled. To enable it one needs to run one simple command after which flashing from Web starts working. (I've tested this.)

This commit updates the website to change the claim into explaining which command is needed.