add bootloader screen sample
Some users have had problems completing this step. This will help to
increase confidence that the user is on the right track.

An image would have been the obvious idea, but using HTML+CSS is more
flexible and works superb even for CLI/TUI browsers (noteworthy for the
users who want to follow the CLI instructions with a minimal web
browser, for example.)