drop key rotation verification instructions
We moved from signify to OpenSSH on 2024-02-14. These instructions were
useful but are consistently causing confusion for people who are missing
that this is an optional step for people who had the signify public key
saved from previous usage. We provide multiple ways to get the current
public key and don't really need this.