Camera version 90 will be in the next release
This didn't end up being included since it was still being tested when
the OS release was initially planned. It has been in Stable a while now
and would have made sense to include, but it will be in the next release
instead. It gets updated out-of-band so it doesn't really matter.