add back important wording to security preview info
The bulletins still get patches regularly added until around a month
away. The monthly ones are usually either empty or close to empty now so
this is mainly relevant to the March 2026 bulletin which doesn't have
most of the patches added yet, since they're mosty going to get added to
it throughout this month.